[refactor]: extract fetch logic to custom hook
