+++ src/views/Episodes/Episodes.js
@@ -1,36 +1,19 @@
 import { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
-import toast from 'react-hot-toast';
 
-import { api } from 'infrastructure';
-
-const { getEpisodes } = api;
+import useFetchApi from 'infrastructure/hooks/useFetchApi';
+import { FETCH_SHOWS_URL } from '../../constants';
 
 const Episodes = () => {
   let params = useParams();
   const [episodes, setEpisodes] = useState([]);
-  const [isLoading, setIsLoading] = useState(false);
-
-  const fetchData = async () => {
-    setIsLoading(true);
-
-    try {
-      const data = await getEpisodes(params.showId);
-      setIsLoading(false);
-      setEpisodes(data);
-    } catch (error) {
-      toast.error(
-        error.response?.data?.message ||
-          error.message ||
-          'Something went wrong!'
-      );
-      setIsLoading(false);
-    }
-  };
+  const [isLoading, data] = useFetchApi(
+    `${FETCH_SHOWS_URL}/${params.showId}/episodes`
+  );
 
   useEffect(() => {
-    fetchData();
-  }, []);
+    setEpisodes(data);
+  }, [data]);
 
   return (
     <main>
