+++ src/views/Shows/index.js
@@ -1,50 +1,31 @@
 import { useEffect, useState, useCallback } from 'react';
-import toast from 'react-hot-toast';
 import debounce from 'just-debounce-it';
-import { api, isTextMatch } from 'infrastructure';
+import { isTextMatch } from 'infrastructure';
+
+import useFetchApi from 'infrastructure/hooks/useFetchApi';
+import { FETCH_SHOWS_URL } from '../../constants';
 
 import ShowsUI from './Shows';
-const { getShows } = api;
 
 function Shows() {
-  const [apiData, setApiData] = useState([]);
-  const [isLoading, setIsLoading] = useState(false);
+  const [shows, setShows] = useState([]);
   const [searchTerm, setSearchTerm] = useState('');
   const [searchResults, setSearchResults] = useState([]);
+  const [isLoading, data] = useFetchApi(FETCH_SHOWS_URL);
 
   const handleChange = (e) => {
     setSearchTerm(e.target.value);
   };
 
-  const fetchData = async () => {
-    setIsLoading(true);
-
-    try {
-      const data = await getShows();
-      setIsLoading(false);
-      setApiData(data);
-      setSearchResults(data);
-    } catch (error) {
-      toast.error(
-        error.response?.data?.message ||
-          error.message ||
-          'Something went wrong!'
-      );
-      setIsLoading(false);
-    }
-  };
-
   const searchFunction = () => {
-    const searchResults = apiData.filter((item) =>
-      isTextMatch(item, searchTerm)
-    );
+    const searchResults = shows.filter((item) => isTextMatch(item, searchTerm));
     setSearchResults(searchResults);
   };
 
   // eslint-disable-next-line react-hooks/exhaustive-deps
   const debouncedSearch = useCallback(
     debounce(() => searchFunction(searchTerm), 500),
-    [apiData, searchTerm]
+    [shows, searchTerm]
   );
 
   useEffect(() => {
@@ -55,14 +36,15 @@ function Shows() {
   }, [debouncedSearch]);
 
   useEffect(() => {
-    fetchData();
-  }, []);
+    setShows(data);
+    setSearchResults(data);
+  }, [data]);
 
   return (
     <>
       <ShowsUI
         searchTerm={searchTerm}
-        apiData={apiData}
+        apiData={shows}
         isLoading={isLoading}
         searchResults={searchResults}
         handleChange={handleChange}
